refact: separate parallel utils matrix_m, add test
